corrigido click box username e link, organizado do Card no arquivo CardStyled
